pyproject.toml
[tool.poetry]
name = "coin_tracker"
version = "0.1.0"
description = "CoinSage - Portfolio Tracker"
authors = ["Oleksandr Hiliazov <oleksandr.hiliazov@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9.10,<3.10"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
sqlmodel = "^0.0.6"
alembic = "^1.7.5"
gunicorn = "^20.1.0"
python-jose = "^3.3.0"
cryptography = "^36.0.1"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
python-dotenv = "^0.19.2"
python-multipart = "^0.0.5"
httpx = "^0.21.3"
greenlet = "^1.1.2"
psycopg2 = "^2.9.3"
SQLAlchemy-Utils = "^0.38.2"
retrying = "^1.3.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
Faker = "^11.3.0"
requests = "^2.27.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
